@@ -1,67 +1,71 @@
 {
-  "name": "zemis",
+  "name": "pickachy",
   "version": "0.0.0",
   "scripts": {
-    "start": "ng serve --host=0.0.0.0 --disable-host-check",
+    "start": "ng serve --host=0.0.0.0 --port 4201 --disable-host-check",
     "lint": "ng lint",
     "lint:fix": "ng lint --fix",
-    "dev:ssr": "ng run zemis:serve-ssr",
-    "serve:ssr": "node dist/zemis/server/main.js",
-    "build:ssr": "ng build && ng run zemis:server && npm run compress",
-    "prerender": "node --max-old-space-size=1500 ./node_modules/@angular/cli/bin/ng run zemis:prerender && npm run compress",
-    "compress": "gzipper c -t 10240 --level=9 --remove-larger dist/zemis/browser"
+    "dev:ssr": "ng run pickachy:serve-ssr",
+    "serve:ssr": "node dist/pickachy/server/main.js",
+    "build:ssr": "ng build && ng run pickachy:server && npm run compress",
+    "prerender": "node --max-old-space-size=1500 ./node_modules/@angular/cli/bin/ng run pickachy:prerender && npm run compress",
+    "compress": "gzipper compress --gzip --brotli --gzip-level=9 --remove-larger --exclude=woff2  dist/pickachy/browser",
+    "source-map-explorer:build": "ng build --source-map",
+    "source-map-explorer:show": "./node_modules/.bin/source-map-explorer dist/pickachy/browser/*",
+    "source-map-explorer:start": "npm run source-map-explorer:build && npm run source-map-explorer:show"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "12.1.0",
-    "@angular/common": "12.1.0",
-    "@angular/compiler": "12.0.5",
-    "@angular/core": "12.0.5",
-    "@angular/forms": "12.0.5",
-    "@angular/platform-browser": "12.1.0",
-    "@angular/platform-browser-dynamic": "12.1.0",
-    "@angular/platform-server": "12.0.5",
-    "@angular/router": "12.0.5",
-    "@angular/service-worker": "12.0.5",
-    "@apollo/client": "3.3.20",
-    "@ng-select/ng-select": "7.0.1",
-    "@nguniversal/express-engine": "12.1.0",
-    "apollo-angular": "2.6.0",
-    "express": "4.17.1",
-    "graphql": "15.5.1",
-    "ngx-progressbar": "6.1.1",
-    "rxjs": "6.6.7",
-    "tslib": "2.3.0",
-    "zone.js": "0.11.4",
-    "sitemap-generator": "8.5.1",
-    "lazysizes": "^5.3.2"
+    "@angular/animations": "13.3.2",
+    "@angular/cdk": "13.3.2",
+    "@angular/common": "13.3.2",
+    "@angular/compiler": "13.3.2",
+    "@angular/core": "13.3.2",
+    "@angular/forms": "13.3.2",
+    "@angular/platform-browser": "13.3.2",
+    "@angular/platform-browser-dynamic": "13.3.2",
+    "@angular/platform-server": "13.3.2",
+    "@angular/router": "13.3.2",
+    "@angular/service-worker": "13.3.2",
+    "@apollo/client": "3.5.10",
+    "@ng-select/ng-select": "8.1.1",
+    "@nguniversal/express-engine": "13.1.0",
+    "ngx-progressbar": "8.0.0",
+    "apollo-angular": "3.0.1",
+    "express": "4.17.3",
+    "graphql": "16.3.0",
+    "lazysizes": "5.3.2",
+    "ngx-toastr": "14.2.3",
+    "rxjs": "7.5.5",
+    "tslib": "2.3.1",
+    "zone.js": "0.11.5"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "12.1.0",
-    "@angular-devkit/architect": "0.1200.5",
-    "@angular-devkit/build-angular": "12.0.5",
-    "@angular-eslint/builder": "12.1.0",
-    "@angular-eslint/eslint-plugin": "12.2.0",
-    "@angular-eslint/eslint-plugin-template": "12.2.0",
-    "@angular-eslint/schematics": "12.1.0",
-    "@angular-eslint/template-parser": "12.2.0",
-    "@angular/cli": "12.0.5",
-    "@angular/compiler-cli": "12.1.0",
-    "@nguniversal/builders": "12.0.2",
-    "@types/express": "^4.17.12",
-    "@types/node": "14.17.4",
-    "@types/sitemap-generator": "^8.5.0",
-    "@typescript-eslint/eslint-plugin": "4.27.0",
-    "@typescript-eslint/parser": "4.27.0",
-    "eslint": "^7.29.0",
-    "eslint-plugin-import": "2.23.4",
-    "eslint-plugin-jsdoc": "35.1.3",
+    "@angular-builders/custom-webpack": "13.1.0",
+    "@angular-devkit/architect": "0.1303.2",
+    "@angular-devkit/build-angular": "13.3.2",
+    "@angular-eslint/builder": "13.2.0",
+    "@angular-eslint/eslint-plugin": "13.2.0",
+    "@angular-eslint/eslint-plugin-template": "13.2.0",
+    "@angular-eslint/schematics": "13.2.0",
+    "@angular-eslint/template-parser": "13.2.0",
+    "@angular/cli": "13.3.2",
+    "@angular/compiler-cli": "13.3.2",
+    "@nguniversal/builders": "13.1.0",
+    "@types/express": "4.17.13",
+    "@types/node": "14.18.12",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "4.31.2",
+    "eslint": "8.13.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsdoc": "35.5.1",
     "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-unused-imports": "1.1.1",
-    "eslint-webpack-plugin": "2.5.4",
-    "ngx-unused-css": "3.0.0",
-    "gzipper": "5.0.0",
-    "prettier": "2.3.2",
-    "typescript": "4.2.3"
+    "eslint-plugin-unused-imports": "2.0.0",
+    "eslint-webpack-plugin": "2.6.0",
+    "gzipper": "7.1.0",
+    "ngx-unused-css": "4.0.0-1",
+    "prettier": "2.6.2",
+    "source-map-explorer": "^2.5.2",
+    "typescript": "4.4.4"
   }
 }
