@@ -1,6 +1,5 @@
 export const environment = {
   production: true,
-  host: 'https://zemis.com.ua',
-  publicApiUrl: 'https://zemis.com.ua/api/v1/graphql',
+  publicApiUrl: 'https://pickachy.com/api/v1/graphql',
   serverApiUrl: 'http://api:5000/api/v1/graphql'
 };
